// src/api/engine-accept-of-day.ts
// tslint:disable
/**
 * Webitel API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 23.07.0
 * Contact: support@webitel.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface EngineAcceptOfDay
 */
export interface EngineAcceptOfDay {
  /**
   *
   * @type {number}
   * @memberof EngineAcceptOfDay
   */
  day?: number
  /**
   *
   * @type {boolean}
   * @memberof EngineAcceptOfDay
   */
  disabled?: boolean
  /**
   *
   * @type {number}
   * @memberof EngineAcceptOfDay
   */
  end_time_of_day?: number
  /**
   *
   * @type {number}
   * @memberof EngineAcceptOfDay
   */
  start_time_of_day?: number
}